Tally counter
Tally counter is developed using java program by using the concept of AWT(Abstract Window Toolkit) in event driven programming.

It will count the number of items by increment the value by 1 when clicking the count button. And the reset button is also available that will reset the count. 

To view the application , run the java code.